fix: fix a wrong assert in receiving hw ack chores: change the test size of default rpc
i assert wrong that there have to be an WR that waits for ACK when receiving a hw ack. however, it is possible that a WR is to big to drain before an ack arrive.
the assert is unneccessary.

+++ simulation/src/point-to-point/model/rdma.h
@@ -122,7 +122,7 @@ using IBVWorkRequest = struct ibv_wr : public SimpleRefCount<ibv_wr> {
     // for RC QP to obtain accurate RTT,
     // in real we do not use this method
     uint64_t send_completion_time;
-    uint32_t last_packet_seq;
+    uint64_t last_packet_seq;
     ibv_wr() { tags.mark_tag_bits = kGeneralTagPayloadBits; }
     ibv_wr(int _mark_tag_bits) { tags.mark_tag_bits = _mark_tag_bits; };
 };
